Исправлены замечания Ревьювера
