+++ README.md
@@ -1,0 +1,12 @@
+# Diplom_1
+
+#Описание
+
+Задание 1: Юнит-тестирование приложения Stellar Burgers.
+
+
+#Используемые технологии
+- Java 11
+- JUnit 5.8.2
+- Maven 3.8.1
+
