+++ src/test/java/praktikum/DatabaseTest.java
@@ -2,16 +2,17 @@ package praktikum;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 
 class DatabaseTest {
     Database database = new Database();
 
     @Test
-    public void databaseIsNotEmpty(){
+    public void databaseIsNotEmpty() {
         assertAll(
-                ()-> assertFalse(database.availableBuns().isEmpty()),
-                ()->assertFalse(database.availableIngredients().isEmpty())
+                () -> assertFalse(database.availableBuns().isEmpty()),
+                () -> assertFalse(database.availableIngredients().isEmpty())
         );
     }
